[qiwen]: mock notification banner update, css update
